requirements.txt
streamlit
tensorflow-cpu==2.15.1
opencv-python-headless
pandas
numpy
scikit-learn
joblib
Pillow